feat: 完成 README 待办 16-34（多脸/渲染/日志/热重启/CameraX 策略等）
参考 CHANGELOG.md：README 待办完成项 16–34。

主要覆盖：
- 多人脸检测与识别输出策略（含节流/稳定性）
- 预览渲染降开销与回退路径（Surface/Bitmap）
- CameraX 分辨率选择迁移与图像分析背压/格式策略校准
- 外部帧输入与 JNI 推帧路径、镜像/翻转语义统一
- 热重启入口与黑屏恢复、自动模式 UI 解释与展示
- App 内日志查看增强：筛选/分页/导出、保留策略与删除
- 文档与验收：UVC 30min 稳定性 runbook；端到端加速可行性报告

验证：本地已通过 Android Debug 组装/单测/lint（gradlew :app:assembleDebug :app:testDebugUnitTest :app:lintDebug）。

+++ src/cpp/src/FaceSearch.cpp
@@ -5,26 +5,30 @@
 
 namespace {
 
+// Optimized to use float instead of double to maximize SIMD vectorization
+// and improve execution speed on the RK3288 (ARM Cortex-A17) platform.
 static float l2Norm(const float* v, std::size_t dim) {
-    double s = 0.0;
+    float s = 0.0f;
     for (std::size_t i = 0; i < dim; i++) {
-        const double x = static_cast<double>(v[i]);
+        const float x = v[i];
         s += x * x;
     }
-    const double n = std::sqrt(s);
-    if (!(n > 0.0)) return 0.0f;
-    return static_cast<float>(n);
+    const float n = std::sqrt(s);
+    if (!(n > 0.0f)) return 0.0f;
+    return n;
 }
 
+// Optimized to use float instead of double to maximize SIMD vectorization
+// and improve execution speed on the RK3288 (ARM Cortex-A17) platform.
 static float cosineSimilarity(const float* a, float aNorm, const float* b, float bNorm, std::size_t dim, bool assumeL2Normalized) {
-    double dot = 0.0;
+    float dot = 0.0f;
     for (std::size_t i = 0; i < dim; i++) {
-        dot += static_cast<double>(a[i]) * static_cast<double>(b[i]);
+        dot += a[i] * b[i];
     }
-    if (assumeL2Normalized) return static_cast<float>(dot);
-    const double denom = static_cast<double>(aNorm) * static_cast<double>(bNorm);
-    if (!(denom > 0.0)) return -1.0f;
-    return static_cast<float>(dot / denom);
+    if (assumeL2Normalized) return dot;
+    const float denom = aNorm * bNorm;
+    if (!(denom > 0.0f)) return -1.0f;
+    return dot / denom;
 }
 
 }  // namespace
